+++ Anarchy/Config/DiscordSocketConfig.cs
@@ -3,8 +3,7 @@
     public class DiscordSocketConfig : DiscordConfig
     {
         public bool ConnectToVoiceChannels { get; set; } = true;
-        public bool GuildSubscriptions { get; set; } = true;
         public bool Cache { get; set; } = true;
-        internal DiscordGatewayIntents Intents { get; set; } = new DiscordGatewayIntents(GatewayIntentCalculator.All());
+        public DiscordGatewayIntent? Intents { get; set; }
     }
 }