+++ Anarchy/REST/Channel/Guild/Voice/VoiceChannel.cs
@@ -38,7 +38,7 @@ namespace Discord
         /// <param name="properties">Options for modifying the channel</param>
         public void Modify(VoiceChannelProperties properties)
         {
-            Update(Client.ModifyChannel(Id, properties));
+            Update(Client.ModifyGuildChannel(Id, properties));
         }