+++ Anarchy/REST/Channel/Private/Group/GroupProperties.cs
@@ -6,8 +6,22 @@ namespace Discord
     /// <summary>
     /// Options for modifying a <see cref="Group"/>
     /// </summary>
-    public class GroupProperties : ChannelProperties
+    public class GroupProperties
     {
+        private readonly DiscordParameter<string> NameProperty = new DiscordParameter<string>();
+        [JsonProperty("name")]
+        public string Name
+        {
+            get { return NameProperty; }
+            set { NameProperty.Value = value; }
+        }
+
+
+        public bool ShouldSerializeName()
+        {
+            return NameProperty.Set;
+        }
+
         internal readonly DiscordParameter<DiscordImage> IconProperty = new DiscordParameter<DiscordImage>();
         [JsonProperty("icon")]
         public DiscordImage Icon
@@ -17,7 +31,7 @@ namespace Discord
         }
 
 
-        public bool ShouldSerialize_icon()
+        public bool ShouldSerializeIcon()
         {
             return IconProperty.Set;
         }